// ui/src/main/angular/ui/modules/lipidcompass-backend-client/model/summarizedLipidDataset.ts
/**
 * LipidCompass Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SummarizedLipidDataset { 
    mzTabResultId?: string;
    quantificationUnitParamName?: string;
    quantificationUnitParamAccession?: string;
    lipidLevel?: string;
    x?: Array<string>;
    y?: Array<number>;
    lipidCategory?: Array<string>;
    lipidClass?: Array<string>;
    groupCount?: Array<number>;
}